Perk system implemented

+++ Assets/Scripts/PlayerHealth.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class PlayerHealth : MonoBehaviour
 {
 
-    public int currentHealth = 100;
+    public float currentHealth = 100;
     public Text healthUI;
     
     // Start is called before the first frame update
